+++ src/lib.rs
@@ -25,10 +25,13 @@
 #![forbid(non_ascii_idents, unsafe_code)]
 #![warn(
     clippy::absolute_paths,
+    clippy::allow_attributes,
+    clippy::allow_attributes_without_reason,
     clippy::as_conversions,
     clippy::as_ptr_cast_mut,
     clippy::assertions_on_result_states,
     clippy::branches_sharing_code,
+    clippy::cfg_not_test,
     clippy::clear_with_drain,
     clippy::clone_on_ref_ptr,
     clippy::collection_is_never_read,
@@ -99,6 +102,7 @@
     clippy::rest_pat_in_fully_bound_structs,
     clippy::same_name_method,
     clippy::semicolon_inside_block,
+    clippy::set_contains_or_insert,
     clippy::shadow_unrelated,
     clippy::significant_drop_in_scrutinee,
     clippy::significant_drop_tightening,
@@ -160,7 +164,7 @@
     variant_size_differences
 )]
 // TODO: Remove on next `derive_more` major version.
-#![allow(clippy::uninlined_format_args)]
+#![expect(clippy::uninlined_format_args, reason = "`derive_more` expansion")]
 
 pub mod cli;
 mod cucumber;