+++ src/step.rs
@@ -206,9 +206,10 @@ impl<World> Collection<World> {
                 }
             };
 
-        // PANIC: Slicing is OK here, as all indices are obtained from the
-        //        source string.
-        #[allow(clippy::string_slice)] // intentional
+        #[expect( // intentional
+            clippy::string_slice,
+            reason = "all indices are obtained from the source string"
+        )]
         let matches = names
             .map(|opt| opt.map(str::to_owned))
             .zip(iter::once(whole_match.as_str().to_owned()).chain(